Rename method related to payment. Update tests.
