--- src/main/java/info/blockchain/wallet/transaction/AdvancedTransaction.java
@@ -1,6 +1,0 @@
-package info.blockchain.wallet.transaction;
-
-public interface AdvancedTransaction extends SimpleTransaction {
-    //maybe add extra from addresses etc (what ever advance transaction will entail)
-    void someExtraMethodForAdvancedTx();
-}
