--- src/test/java/info/blockchain/wallet/transaction/AdvancedTransactionTest.java
@@ -1,14 +1,0 @@
-package info.blockchain.wallet.transaction;
-
-import org.junit.Test;
-
-public class AdvancedTransactionTest {
-
-    @Test
-    public void someExtraMethodForAdvancedTxTest() {
-
-        //Todo - advance transaction
-        AdvancedTransaction advTransaction = new BaseTransaction();
-        advTransaction.someExtraMethodForAdvancedTx();
-    }
-}
